Money Format

+++ Assets/Scripts/Resources/ResourceBank.cs
@@ -1,7 +1,7 @@
 public class ResourceBank
 {
-    private int _money;
-    public int Money { get => _money; set => _money = value; }
+    private long _money;
+    public long Money { get => _money; set => _money = value; }
 
     public ResourceBank(int money)
     {
